Separate HttpContent into separate class from Buffer.
